Maintenance, Bugfixes #CHANGE: Former Duality.Resources.Pixmap.Layer class is now Duality.Drawing.PixelData. #FIX: Fixed Texture and Material creation in editor not using the correct file endings anymore.
