+++ Duality/Drawing/Canvas.cs
@@ -285,7 +285,7 @@ namespace Duality.Drawing
 			ContentRef<Texture> dashTexRef = new ContentRef<Texture>(null, dashTextPath);
 			if (!dashTexRef.IsAvailable)
 			{
-				Pixmap.Layer pxLayerDash = new Pixmap.Layer(32, 1);
+				PixelData pxLayerDash = new PixelData(32, 1);
 				for (int i = 31; i >= 0; i--) pxLayerDash[i, 0] = ((patternBits & (1U << i)) != 0) ? ColorRgba.White : ColorRgba.TransparentWhite;
 				Pixmap pxDash = new Pixmap(pxLayerDash);
 				Texture texDash = new Texture(pxDash, TextureSizeMode.Stretch, TextureMagFilter.Nearest, TextureMinFilter.Nearest, TextureWrapMode.Repeat);
