+++ Duality/Duality.csproj
@@ -116,6 +116,8 @@
     <Compile Include="Drawing\Enums\ClearFlag.cs" />
     <Compile Include="Drawing\Enums\ColorDataElementType.cs" />
     <Compile Include="Drawing\Enums\ColorDataLayout.cs" />
+    <Compile Include="Drawing\Enums\ImageScaleFilter.cs" />
+    <Compile Include="Drawing\PixelData.cs" />
     <Compile Include="Resources\ShaderInfo\ShaderType.cs" />
     <Compile Include="Drawing\Enums\RenderMatrix.cs" />
     <Compile Include="Drawing\Enums\PerspectiveMode.cs" />
